requirements.txt
discord-py-interactions
discord
python-dotenv